refactor: Extract VerticalWhisker to its own file
